BPP: Pull Entries and Entry as first classes out of BuildProperties
