+++ BuildPropertiesPlugin/plugin/src/main/groovy/com/novoda/buildproperties/BuildProperties.groovy
@@ -45,56 +45,4 @@ class BuildProperties {
     }
   }
 
-  static abstract class Entries {
-
-    abstract boolean contains(String key)
-
-    protected abstract Object getValueAt(String key)
-
-    Entry getAt(String key) {
-      new Entry(key, {
-        getValueAt(key)
-      })
-    }
-
-    abstract File getParentFile()
-
-    abstract Enumeration<String> getKeys()
-
-  }
-
-  static class Entry {
-    private final String key
-    private final Closure<Object> value
-
-    Entry(String key, Closure<Object> value) {
-      this.key = key
-      this.value = value
-    }
-
-    String getKey() {
-      key
-    }
-
-    Boolean getBoolean() {
-      Boolean.parseBoolean(string)
-    }
-
-    Integer getInt() {
-      value.call() as Integer
-    }
-
-    Double getDouble() {
-      value.call() as Double
-    }
-
-    String getString() {
-      value.call() as String
-    }
-
-    Object getValue() {
-      value.call()
-    }
-  }
-
 }
